requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.43.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
aws-cdk.aws-apigatewayv2-alpha==2.43.0.a0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha==2.43.0a0
    # via -r requirements.in
aws-cdk.aws-lambda-python-alpha==2.43.0a0
    # via -r requirements.in
cattrs==22.1.0
    # via jsii
constructs==10.1.109
    # via
    #   -r requirements.in
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
exceptiongroup==1.0.0rc9
    # via cattrs
jsii==1.67.0
    # via
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via
    #   -c backend/api/runtime/requirements.txt
    #   jsii
six==1.16.0
    # via
    #   -c backend/api/runtime/requirements.txt
    #   python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   constructs
    #   jsii
typing-extensions==4.3.0
    # via
    #   cattrs
    #   jsii
